Create Longest subsequence such that difference between adjacents is 1
The problem “Longest subsequence such that difference between adjacents is one” states that you are given an integer array. Now you need to find the length of longest subsequence such that the difference of adjacent elements is 1.
Example: [1,2,3,4,7,5,8,4]
As we can see that there are two subsequences that follow the condition. They are {1, 2, 3, 4, 5, 4} and the other one is {7, 8}. So the longest subsequence is the first one. Thus the answer is 6.